Switch to the new development approach and to new version of the FusionInventory-Agent project.
With the new series 2.3.x of the FusionInventory-Agent project some things have
changed, mainly two:

 - Now there is only a source code (and a repository) and not four like before.

   The agent and all its tasks are in only one package, the FusionInventory
   Agent; the other projects related to its tasks (FusionInventory Agent Task
   Deploy, FusionInventory Agent Task ESX and FusionInventory Agent Task
   Network) are desprecated now (I suppose).

 - Now the development approach is different. Read the following link for more
   information

   http://www.fusioninventory.org/news/2013/02/20/agent_development_process_changes/

   At the moment, there is not a 2.3.x branch. The new FusionInventory Agent is
   in development into the main branch.

              +- c <- d <- e   # 2.2.x branch
             /
       a <- b <- f <- g <- h   # master branch

These changes into the FusionInventory Agent project have some implications in
the FusionIventory Agent Windows Installer project. All these changes prepare
the FusionInventory Agent Windows Installer project to build the Microsoft
Windows installer for the, now in development, FusionInventory Agent 2.3.x.

Currently, this project is prepared to build the Microsoft Windows installer for
the current state of the master branch of the FusionInventory-Agent project.

Nowadays, NSIS works (compiles) perfectly but the FusionInventory-Agent for
Windows has execution problems. I hope most of them are related with the
development phase of the FusionInventory-Agent project and not with this
project.

I hope that it is easier now to check the development process on Microsoft
Windows systems.