Make possible the construction of two types of installers.
Now, making use of the new constant FIAI_RELEASE_TYPE is possible to construct
two different installers, one for the stable releases and other one for the
development releases.

It's important to understand that a stable version of the installer can install
a stable or a development release of the agent.

In the same way, a development version of the installer can install a
development or a stable release of the agent.