Temporarily add directory "%LOCALAPPDATA%\Pandoc" to PATH.
With the latest versions of Pandoc you can install it per-user (default)
or per-machine. If it's installed per-user, Pandoc is located into
directory "%LOCALAPPDATA%\Pandoc". If it's installed per-machine, Pandoc
is located into directory "%PROGRAMFILES%\Pandoc" (in systems of 32bit) or
into directory "%PROGRAMFILES(X86)%\Pandoc" (in systems of 64bit).